requirements.txt
torch
pandas
python-telegram-bot 